+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fix `Pledge.done()` resolving with object versions of primitives
 * Make `Blast.require(path)` use `vm` module instead of monkey-patching `require`
 * Add `Cache#clear()` method
+* Add `Function.postInherit(task)`, which will execute the task after the class has been inherited
 
 ## 0.8.18 (2024-01-19)